requirements.txt
python-dotenv
livekit-agents~=1.4
livekit-plugins-silero
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-turn-detector
livekit-plugins-noise-cancellation
